+++ src/main/java/sopt/org/hmh/domain/auth/dto/response/LoginResponse.java
@@ -1,7 +1,7 @@
 package sopt.org.hmh.domain.auth.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import sopt.org.hmh.global.auth.jwt.TokenResponse;
+import sopt.org.hmh.global.auth.jwt.dto.TokenResponse;
 
 public record LoginResponse(
         Long userId,
