+++ src/main/java/sopt/org/hmh/global/config/SecurityConfig.java
@@ -10,8 +10,8 @@ import org.springframework.security.config.annotation.web.configurers.AbstractHt
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import sopt.org.hmh.global.auth.jwt.JwtProvider;
-import sopt.org.hmh.global.auth.jwt.JwtValidator;
+import sopt.org.hmh.global.auth.jwt.service.JwtProvider;
+import sopt.org.hmh.global.auth.jwt.service.JwtValidator;
 import sopt.org.hmh.global.auth.security.JwtAuthenticationEntryPoint;
 import sopt.org.hmh.global.auth.security.JwtAuthenticationFilter;
 import sopt.org.hmh.global.auth.security.exception.ExceptionHandlerFilter;
